Remove rerun-if-changed for files in the package
Cargo normally doesn't pay attention to the file modification times in crates pulled from a registry since they should always stay unmodified. This allows Cargo to avoid rebuilding dependencies when e.g. you delete the Cargo cache and it has to download dependencies again.

However, this was not true for jemalloc-sys, which explicitly asked to be rerun on changes to the vendored jemalloc source. I imagine this was originally added to ease testing while working on the crate itself, but it should be removed to better work in the more common use case of pulling it from crates.io.

Signed-off-by: Steven Fackler <sfackler@gmail.com>