Attempt supporting spaces in CC/SHELL path
Autoconf does not support spaces in the compiler path. This is not
usually a problem, as on most platforms, the compiler will reside in a
path without spaces (such as /usr/bin or /opt). However, on windows, the
MSVC compiler is usually found under C:\Program Files, which causes all
hell to break loose.

Similarly, the jemalloc Makefile also assumes the shell (stored in the
SHELL variable) is in a folder without spaces, and is used unquoted all
throughout the makefile, causing issues if the shell is in C:\Program
Files, such as when using Git Bash (which installs in C:\Program
Files\Git\bin\bash).

Thankfully, windows has something called "Short Paths", which we can use
to turn C:\Program Files into C:\PROGRA~2, eliminating the path.

If the build script detects the compiler path has a space, and is
running on windows, it will attempt to turn the path into a short path
and run with this.

Signed-off-by: roblabla <unfiltered@roblab.la>