Set cc-rs env vars when invoking MSVC CC
When invoking MSVC CC, a handful of environment variables must be set
for the compiler to find its supporting files (such as system includes,
libraries, etc...). When invoking it through cc-rs, these environment
variables are automatically detected and set.

However, we currently don't set those environment variables when running
`./configure` or `make`, leading to those commands failing due to the
compiler not having access to any of the system include/libraries.

To fix this, we ask cc-rs to give us the environment variables to set,
and set them whenever running `configure` and `make`.

Signed-off-by: roblabla <unfiltered@roblab.la>