Disable background_threads support on windows
Background Threads are unsupported on windows in jemalloc, due to
requiring pthreads (which aren't available on windows, at least on the
msvc target). Trying to enable it will cause a runtime error causing the
program to immediately abort.

As such, to avoid problems, we set windows as a target that does not
support background_threads.

Signed-off-by: roblabla <unfiltered@roblab.la>